requirements.txt
streamlit
pandas
qrcode
pillow
openpyxl
streamlit-extras
opencv-python